+++ pom.xml
@@ -101,14 +101,6 @@
         <!--+
             | Test scope
             +-->
-       <!-- https://mvnrepository.com/artifact/org.apache.httpcomponents/httpclient -->
-        <dependency>
-            <groupId>org.apache.httpcomponents</groupId>
-            <artifactId>httpclient</artifactId>
-            <version>4.5.2</version>
-            <scope>test</scope>
-        </dependency>
-
         <dependency>
             <groupId>org.apache.camel</groupId>
             <artifactId>camel-testng</artifactId>
@@ -121,7 +113,6 @@
             <version>${org.testng.version}</version>
             <scope>test</scope>
         </dependency>
-
     </dependencies>
 
     <!--+
